+++ Cargo.toml
@@ -1,16 +1,14 @@
 [package]
-name = "day2"
+name = "todo"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-
 [dependencies.rocket]
 version = "0.5.0-rc.2"
-features = ["json"]
+features = ["json", "uuid"]
 
-[dependencies]
-sysinfo = "0.25.3"
-uptime_lib = "0.2.2"
-chrono = "0.4.22"
+[dependencies.uuid]
+version = "1"
+features = ["v1", "v4"]
