--- day1/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "day1"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ferris-says = "0.2"
