--- day3/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "day3"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies.rocket]
-version = "0.5.0-rc.2"
-features = ["json", "uuid"]
-
-[dependencies.uuid]
-version = "1"
-features = ["v1", "v4"]
